[rn][win][Fabric] Implement inverted prop in ScrollView
Summary:
In order to support native inversion, we need to use the inverted prop
to implement functionality like anchoring and restartable scroll-to-end.

This adds support for the inverted prop in a similar way to how we
implemented it in Paper. The main difference is the frequency at which
the InvalidateArrange method is called. In Paper, we know whenever a
Yoga node is marked with "hasNewLayout", SetLayoutProps will be called
on it's ScrollView ancestor's native ViewManager. In Fabric, we do not
have similar guarantees. The closest behavior I can come up with is to
call InvalidateArrange whenever the content container size changes.
T143666093 tracks a fix for ensuring InvalidateArrange is called often
enough.

Test Plan:
Changed FlatList-basic.js RNTester example to use inverted list by default:
https://pxl.cl/2rnK5


Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42787560

Tasks: T140460683