[rn][win][Fabric] Adds scrollTo* commands to ScrollView
Summary:
Implements the scrollTo and scrollToEnd commands for ScrollView in
Fabric. The implementation is equivalent to what is currently committed
in the react-native-windows open source project. Once we get the native
inverted behaviors working, this command will change a bit to use the
ScrollViewViewChanger implementation to restart `scrollToEnd` in case
layout interrupts the scroll animation.

Test Plan:
scrollTo command works:
https://pxl.cl/2rnLb

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42805829

Tasks: T138848828