[rn][win][Fabric] Fix crash on null View backgroundColor
Summary:
Previously, when an animation occurred and a backgroundColor was reset
from some valid color to null, the app would crash due to a spurious
dereference of the backgroundColor value. This was not needed. We can
also directly set the backgroundColor to Panel::Background instead of
ViewPanel::ViewBackground now that we depend on the Grid-based ViewPanel
implementation.

Test Plan: Clicking an item in FlatList-basic.js example in RNTester previously crashed, no longer crashes.

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42888813