[rn][win][Fabric] Only set transform if not identity
Summary:
We previously updated the transform on all views on initial render, even
when the transform was an identity transform. It's rather expensive to
set up transforms in react-native-windows, as we use composition
property sets for transform values. This change ensures that we only
set up the composition animation for transforms the first time the
transform value is not an identity matrix.

Test Plan:
Tested using PerfView, we use 10% fewer CPU cycles on render.

Also, JS driven transform animations still work as expected:
https://pxl.cl/2rqcg

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42898751