[rn][win][Fabric] Only keep event emitters for isPressable inlines
Summary:
We spend a non-trivial amount of time setting up pressable event
handling for virtual spans of Text when rendering in Fabric. Most Text
components do not support press events, so this is "wasted" work. Even
more problematically, we perform a search over each raw text span on
mouse move events, which is a waste of CPU cycles.

This change revives the isPressable prop used in Paper for use in
Fabric, so only the runs that exist in a pressable context will be
searched as a potential hit test target and fire events.

Test Plan: https://pxl.cl/2rvPS

Reviewers: shawndempsey, lefever, helenistic

Reviewed By: lefever, helenistic

Differential Revision: https://phabricator.intern.facebook.com/D42926441

Tasks: T140425180