[rn][win][Fabric] Ensure transparent brush for hoverables
Summary:
In Paper for react-native-windows, we set a transparent background brush
if either onMouseEnter or onMouseLeave events are subscribed. This
changes the ViewComponentView to do the same for Fabric.

Test Plan:
Previously, hovering in the transparent area of the bordered outer view failed the hit test (because Windows views are not hit testable unless they have a background brush). Now:
https://pxl.cl/2rF2f

Reviewers: shawndempsey, lefever, helenistic

Reviewed By: helenistic

Differential Revision: https://phabricator.intern.facebook.com/D42965098

Tasks: T144117065