[rn][win][Fabric] Remove workaround for packaged asset flag
Summary:
Prior to PR #10621, the ImageViewManager for Paper and
ImageViewComponent for Fabric required a workaround where file:// URIs
were remapped to an absolute path using the InstanceSettings from the
ReactContext.

After #10621, the URI is already fully resolved by the time it reaches
the native renderer for Image, so we do not need to do any URI rewrites.

Test Plan:
Images still load in RNTester:
{F868525679}

Reviewers: shawndempsey, helenistic, chpurrer, lefever

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D43159275

Tasks: T140424672