[rn][win][Fabric] Fix minor issue where ScrollView borders do not render
Summary:
We need to mark that the ScrollView needs a border update in
updateLayoutMetrics as we do in ViewComponentView. This change also
includes a header that was previously included from another header, but
it's better to be explicit that this header is needed.

Test Plan:
We missed some border updates before...
|Before|After|
| {F868674176} | {F868675426} |

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D43160624