[rn][win][Fabric] Fix resizeMode="stretch" behavior on Image
Summary:
As it turns out, "stretch" is the default value for ImageProps in
Fabric. We were missing the prop update when "stretch" was set because
there is no delta from the default value.

To solve this, we set the default prop value for ResizeMode when the
ImageComponentView is constructed.

Test Plan: {F868614933}

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: helenistic

Differential Revision: https://phabricator.intern.facebook.com/D43161449

Tasks: T145215774