[PR] Adds collapsable prop to Paper VM for View (#11249)
Summary:
Fabric reads native view configs from the Paper UIManager. In order for
the `collapsable` prop to flow from JS to the native props for Fabric,
we need to have this prop set in the VM configuration for View in Paper.

Test Plan: Verified that collapsable prop is set on ViewProps after this change.

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D43465229