[rn][win][Fabric] Initial implementation of LegacyABIViewComponentView
Summary:
This is a basic implementation of LegacyABIViewComponentView. It adds
the component descriptor, state, props, event emitter, etc. for
connecting a Fabric component view to the ABI-safe IViewManagers used in
react-native-windows Paper. It does not yet wire up functionality like
`IViewManagerRequiresNativeLayout`. It does, however, serialize props,
mount and unmount children, and create the appropriate XAML view using
the IViewManager.

Test Plan:
Attempt to render `nativeLayout.tsx`, which uses two custom IViewManager implementations in Playground-win32...

The layout is not correct, due to missing functionality for IViewManagerRequiresNativeLayout, but this at least demonstrates that children are mounted correctly and props are serialized and sent to the IViewManager:

{F879953107}

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: lefever

Differential Revision: https://phabricator.intern.facebook.com/D43463895

Tasks: T138599471