[rn][win][Fabric] Wire up events for LegacyABIViews
Summary:
ABI view managers use the `IReactContext::DispatchEvent` method to
dispatch events to JavaScript. `IReactContext` routes dispatched events
to XamlUIService. This change modifies XamlUIService to first check if
the tag for the provided view is managed by FabricUIManager, and if so,
checks if the component view has a LegacyABIViewEventEmitter that can be
used to dispatch the event.

Test Plan:
Wired up custom event for click on GridItemViewManager and fire an alert when called:
https://pxl.cl/2tvX7

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: lefever

Differential Revision: https://phabricator.intern.facebook.com/D43474369

Tasks: T146189593