[rn][win][Fabric] Enable fast text optimizations
Summary:
XAML has functionality where it uses optimized text layout when the Text
property is used on TextBlock rather than its Inlines property. This is
functionality that already exists in the Paper UIManager, but the
ParagraphComponentView for Fabric was not set up to leverage this
optimization. This change keeps most of the code for building the inline
collection, but while building, determines whether the text is eligible
for optimization, and if so, clears the Inline collection and sets the
Text property.

This change also fixes numerous bugs in Text:
- Avoids setting a FontWeight if no font weight is set in ParagraphProps
- Avoids setting a FontStyle if no font style is set in ParagraphProps
- Implements textDecorationLineType functionality on the Paragraph node
- Fixes an issue where updating the backgroundColor on the Paragraph
node did not always work correctly.
- Limits each fragment to at most one Span (rather than nesting one per
use of accessibilityRole="link", fontStyle, and textDecorationLineType).
- Implements textDecorationLineType="strikethrough"

This change generally results in more cases where text can be optimized
than was possible in Paper (largely due to virtual text node flattening
in AttributedString).

Test Plan:
I turned on DebugSettings::IsTextPerformanceVisualizationEnabled, which turns optimized text green. Not shown here, but there's a lot more optimized text in Fabric than in Paper :)

 https://pxl.cl/2tMGr

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: lefever

Subscribers: taskcreeper

Differential Revision: https://phabricator.intern.facebook.com/D43539443

Tasks: T142895298, T138849510