[rn][win][Fabric] Mitigate XAML Islands TextCommandBarFlyout bug
Summary:
In XAML Islands, there is a singleton TextCommandBarFlyout that is used
by default in selectable TextBlocks, TextBox, and PasswordBox. This
singleton is attached to a single XamlRoot, and will only render on the
first window it is opened on.

To work around this, we assign a unique TextCommandBarFlyout to every
TextBox and PasswordBox, as well as to every selectable TextBlock.

Test Plan:
Context menus for Text and TextInput still work as expected:
https://pxl.cl/2tP50

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: lefever

Differential Revision: https://phabricator.intern.facebook.com/D43548049

Tasks: T146414148