[rn][win][Fabric] Fix up eventCount behavior in TextInput
Summary:
As a controlled component, TextInput needs to be sure that updates from
JS don't clobber recent updates from the native thread. To do this, an
eventCount is sent to JavaScript with each onChange event. This
eventCount is sent back to native when controlled updates occur, and we
only modify the native component if the eventCount matches the
mostRecentEventCount in the native component state.

There were a few bugs with this logic, largely derived from the fact
that we had two separate counters: m_nativeEventCount and
m_mostRecentEventCount. This change removes the redundant
m_nativeEventCount and ensures that the m_mostRecentEventCount is
updated on native changes.

Test Plan: 

Reviewers: 

Subscribers: 

Tasks: 

Tags: 


Differential Revision: https://phabricator.intern.facebook.com/D43561633