[rn][win][Fabric] Fix event order for onChange and onSelectionChange
Summary:
On other platforms, onSelectionChange fires after onChange. This change
ensures that we follow the same pattern on Fabric. We will emit the
selection change event synchronously with the TextChanging callback when
SelectionChanging was called prior to TextChanging. Otherwise, we'll
emit the selection change event in SelectionChanged. This is likely a
minor optimization - we could always emit the event from
SelectionChanged, but firing both the onChange and onSelectionChange
event in the same batch of events may reduce the number of re-renders
for controlled components that mutate state on both events.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: https://phabricator.intern.facebook.com/D43609944